+++ fsl/wrappers/fslio.py
@@ -10,17 +10,19 @@ These scripts are actually implemented within fslpy, so the wrapper functions
 invoke the relevant code directly, rather than invoking a separate process.
 """
 
-from fsl.scripts import (imcp   as imcp_script,
-                         imglob as imglob_script,
-                         imln   as imln_script,
-                         immv   as immv_script,
-                         imrm   as imrm_script,
-                         imtest as imtest_script)
+from fsl.scripts import (imcp       as imcp_script,
+                         imglob     as imglob_script,
+                         imln       as imln_script,
+                         immv       as immv_script,
+                         imrm       as imrm_script,
+                         imtest     as imtest_script,
+                         remove_ext as remove_ext_script,
+                         tmpnam     as tmpnam_script)
 
 
-def imcp(*args):
+def imcp(*paths):
     """Wrapper for the ``imcp`` script. """
-    imcp_script.main(args)
+    imcp_script.imcp(*paths)
 
 
 def imglob(*images, extension=False, extensions=False):
@@ -35,19 +37,19 @@ def imglob(*images, extension=False, extensions=False):
     return imglob_script.imglob(images, output)
 
 
-def imln(*args):
+def imln(target, linkbase):
     """Wrapper for the ``imln`` script. """
-    imln_script.main(args)
+    imln_script.imln(target, linkbase)
 
 
-def immv(*args):
+def immv(*paths):
     """Wrapper for the ``immv`` script. """
-    immv_script.main(args)
+    immv_script.immv(*paths)
 
 
-def imrm(*args):
+def imrm(*paths):
     """Wrapper for the ``imrm`` script. """
-    imrm_script.main(args)
+    imrm_script.imrm(*paths)
 
 
 def imtest(path):
@@ -55,3 +57,27 @@ def imtest(path):
     exists, ``False`` otherwise.
     """
     return imtest_script.imtest(path)
+
+
+def remove_ext(*paths):
+    """Wrapper for the ``remove_ext`` script.
+
+    Returns the specified paths with extensions removed. If one path is
+    provided, a string is returned. Otherwise (multiple paths provided), a list
+    is returned.
+    """
+    paths = remove_ext_script.remove_ext(*paths)
+
+    if len(paths) == 1: return paths[0]
+    else:               return paths
+
+
+def tmpnam(path=None):
+    """Wrapper for the ``tmpnam`` script. """
+    return tmpnam_script.tmpnam(path)
+
+
+def fslchfiletype(fmt, oldfile, newfile=None):
+    """Wrapper for the ``fslchfiletype`` script. """
+    from fsl.scripts import fslchfiletype as fslchfiletype_script
+    fslchfiletype_script.fslchfiletype(fmt, oldfile, newfile)